refactor: split pl-mcp-server tools into separate files, fix bugs
> "Refactor server.ts based on PR review comments"

- Split monolithic server.ts into 11 tool files under src/tools/
- Add ToolContext interface and textResult helper in tools/types.ts
- Fix race condition: set transport in map BEFORE handleRequest()
- Add try-catch wrapper around HTTP request handler body
- Fix non-null assertion in EADDRINUSE retry loop (use local variable)
- Fix import extensions in index.ts (remove .js, match Bundler convention)
- Fix Node engine version: >=22.19.0 → >=22.0.0
