+++ lib/node/pl-mcp-server/package.json
@@ -47,6 +47,6 @@
     "@milaboratories/pl-middle-layer": ">=1.48.0"
   },
   "engines": {
-    "node": ">=22.19.0"
+    "node": ">=22.0.0"
   }
 }
